Move konnectivity server out of KAS pod
Reverts the change from
https://github.com/openshift/hypershift/pull/2942 and puts
konnectivity-server back in its own deployment.
