+++ control-plane-operator/controllers/hostedcontrolplane/hostedcontrolplane_controller.go
@@ -1578,7 +1578,7 @@ func (r *HostedControlPlaneReconciler) reconcileKonnectivityServerService(ctx co
 	}
 	konnectivityServerService := manifests.KonnectivityServerService(hcp.Namespace)
 	if _, err := createOrUpdate(ctx, r.Client, konnectivityServerService, func() error {
-		return kas.ReconcileKonnectivityServerService(konnectivityServerService, p.OwnerRef, serviceStrategy)
+		return konnectivity.ReconcileServerService(konnectivityServerService, p.OwnerRef, serviceStrategy)
 	}); err != nil {
 		return fmt.Errorf("failed to reconcile Konnectivity service: %w", err)
 	}
@@ -1588,7 +1588,7 @@ func (r *HostedControlPlaneReconciler) reconcileKonnectivityServerService(ctx co
 	konnectivityRoute := manifests.KonnectivityServerRoute(hcp.Namespace)
 	if util.IsPrivateHCP(hcp) {
 		if _, err := createOrUpdate(ctx, r.Client, konnectivityRoute, func() error {
-			return kas.ReconcileKonnectivityInternalRoute(konnectivityRoute, p.OwnerRef)
+			return konnectivity.ReconcileInternalRoute(konnectivityRoute, p.OwnerRef)
 		}); err != nil {
 			return fmt.Errorf("failed to reconcile Konnectivity server internal route: %w", err)
 		}
@@ -1598,7 +1598,7 @@ func (r *HostedControlPlaneReconciler) reconcileKonnectivityServerService(ctx co
 			if serviceStrategy.Route != nil {
 				hostname = serviceStrategy.Route.Hostname
 			}
-			return kas.ReconcileKonnectivityExternalRoute(konnectivityRoute, p.OwnerRef, hostname, r.DefaultIngressDomain)
+			return konnectivity.ReconcileExternalRoute(konnectivityRoute, p.OwnerRef, hostname, r.DefaultIngressDomain)
 		}); err != nil {
 			return fmt.Errorf("failed to reconcile Konnectivity server external route: %w", err)
 		}
@@ -1947,7 +1947,7 @@ func (r *HostedControlPlaneReconciler) reconcileKonnectivityServiceStatus(ctx co
 			return
 		}
 	}
-	return kas.ReconcileKonnectivityServerServiceStatus(svc, route, serviceStrategy, events.NewMessageCollector(ctx, r.Client))
+	return konnectivity.ReconcileServerServiceStatus(svc, route, serviceStrategy, events.NewMessageCollector(ctx, r.Client))
 }
 
 func (r *HostedControlPlaneReconciler) reconcileOAuthServiceStatus(ctx context.Context, hcp *hyperv1.HostedControlPlane) (host string, port int32, message string, err error) {
@@ -2563,13 +2563,14 @@ func (r *HostedControlPlaneReconciler) reconcileKonnectivity(ctx context.Context
 	r.Log.Info("Reconciling Konnectivity")
 	p := konnectivity.NewKonnectivityParams(hcp, releaseImageProvider, infraStatus.KonnectivityHost, infraStatus.KonnectivityPort, r.SetDefaultSecurityContext)
 	serverDeployment := manifests.KonnectivityServerDeployment(hcp.Namespace)
-	// Remove the konnectivity-server deployment if it exists
-	if _, err := util.DeleteIfNeeded(ctx, r, serverDeployment); err != nil {
-		return fmt.Errorf("failed to remove konnectivity-server deployment: %w", err)
+	if _, err := createOrUpdate(ctx, r, serverDeployment, func() error {
+		return konnectivity.ReconcileServerDeployment(serverDeployment, p.OwnerRef, p.ServerDeploymentConfig, p.KonnectivityServerImage)
+	}); err != nil {
+		return fmt.Errorf("failed to reconcile konnectivity server deployment: %w", err)
 	}
 	serverLocalService := manifests.KonnectivityServerLocalService(hcp.Namespace)
 	if _, err := createOrUpdate(ctx, r, serverLocalService, func() error {
-		return kas.ReconcileKonnectivityServerLocalService(serverLocalService, p.OwnerRef)
+		return konnectivity.ReconcileServerLocalService(serverLocalService, p.OwnerRef)
 	}); err != nil {
 		return fmt.Errorf("failed to reconcile konnectivity server local service: %w", err)
 	}
